Add support for alternative Storage implementations.
Currently, hc uses the file system to store session keys. Internally,
this is acutally implemented as a generic Storage interface, however
it is currently not possible for the users to use an alternative
implementation if they so desire.

This change introduces a new config option, Config.Storage. This
optional option lets users provide a custom Storage implementation. In
my case, for example, I have an internal implementation that talks
to a relational database.